# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShiguReader API",
    "description": "ShiguReader backend API — 类似 FastAPI 的交互式文档",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "HistoryItem": {
        "title": "HistoryItem",
        "type": "object",
        "required": [
          "filepath",
          "opened_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "filepath": {
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "opened_at": {
            "type": "integer"
          }
        }
      },
...
      "LocalCheckHit": {
        "title": "LocalCheckHit",
        "type": "object",
        "required": [
          "name",
          "path",
          "thumbnail_url",
          "match_level",
          "confidence"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "downloaded"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "likely"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "same_author"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "different"
                ]
              }
            ]
          },
          "confidence": {
            "type": "number"
          }
        }
      },
      "LocalCheckResult": {
        "title": "LocalCheckResult",
        "type": "object",
        "required": [
          "q",
          "match_level",
          "confidence",
          "reason",
          "hits"
        ],
        "properties": {
          "q": {
            "type": "string"
          },
          "match_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "downloaded"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "likely"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "same_author"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "different"
                ]
              }
            ]
          },
          "confidence": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalCheckHit"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/search/local-check-batch": {
      "post": {
        "operationId": "localCheckBatch",
        "summary": "批量本地持有检查（油猴脚本用）",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "LocalCheckBatchRequest",
                "type": "object",
                "properties": {
                  "queries": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "limit": {
                    "default": 5,
                    "type": "integer"
                  },
                  "presence_filter": {
                    "default": "all",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "all"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "watched"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "scanned_recent"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "LocalCheckBatchResponse",
                  "type": "object",
                  "required": [
                    "results"
                  ],
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocalCheckResult"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/list": {
      "get": {
        "operationId": "listHistory",
        "summary": "获取阅读历史列表（分页）",
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "page_size",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort_order",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "HistoryListResponse",
                  "type": "object",
                  "required": [
                    "items",
                    "page",
                    "page_size",
                    "total",
                    "total_pages"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HistoryItem"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "total_pages": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/record": {
      "post": {
        "operationId": "recordHistory",
        "summary": "记录一次阅读",
        "tags": [
          "History"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "HistoryRecordRequest",
                "type": "object",
                "required": [
                  "filepath"
                ],
                "properties": {
                  "filepath": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "HistoryRecordResponse",
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ok"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "skipped"
                          ]
                        }
                      ]
                    },
                    "reason": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}